Implemented conditions for target hole detection
The calculated distance to object must be within 2 metres, the object must be reasonably centered in the frame and the viewing angle must be less than +/-10 degrees.
Added some code to almost guarantee findTargetHoles function returns 2 rectangular contours. There was an issue that if either of the 2 largest contours was not a rectangle then the function would return nothing at all. Therefore, if the target holes were not amongst the largest contours they would not be recognised at all.