Refactored the Color System class
Extracted some key sections of code into functions in preparation for combining the RectangleVision and ColorSystem classes in order to make the application run as a single unit rather than two separate.
Reworked the sequence recognition code in order to improve its stability and reliability in the event of background noise being detected, missed colours etc.